+++ src/sn/GrpLayer.ts
@@ -10,7 +10,7 @@ import {Layer} from './Layer';
 import {CmnLib, int, IEvtMng, argChk_Boolean, argChk_Num, getFn} from './CmnLib';
 import {HArg} from './Grammar';
 import {IMain, IVariable, SYS_DEC_RET} from './CmnInterface';
-import {Config, SEARCH_PATH_ARG_EXT} from './Config';
+import {Config} from './Config';
 import {SysBase} from './SysBase';
 import {Sprite, Container, Texture, BLEND_MODES, utils, Loader, LoaderResource, AnimatedSprite, Rectangle, RenderTexture, Application} from 'pixi.js';
 import {EventListenerCtn} from './EventListenerCtn';
@@ -18,6 +18,7 @@ import {SoundMng} from './SoundMng';
 import {IMakeDesignCast} from './LayerMng';
 import {GrpLayDesignCast} from './DesignCast';
 import {DebugMng} from './DebugMng';
+import {SEARCH_PATH_ARG_EXT} from './ConfigBase';
 
 export type IFncCompSpr = (sp: Sprite)=> void;
 
@@ -35,6 +36,7 @@ interface IResAniSpr {
 		animationSpeed? :number;
 	};
 }
+
 
 export class GrpLayer extends Layer {
 	static	readonly	#elc	= new EventListenerCtn;
@@ -195,7 +197,7 @@ export class GrpLayer extends Layer {
 //			GrpLayer.ldrHFn[f.fn] = 1;
 
 			needLoad = true;
-			const url = GrpLayer.#cfg.searchPath(f.fn, SEARCH_PATH_ARG_EXT.SPRITE);
+			const url = GrpLayer.#cfg.searchPath(f.fn, SEARCH_PATH_ARG_EXT.SP_GSM);
 			const xt = this.#sys.crypto
 			? {xhrType: (url.slice(-5) === '.json')
 				? LoaderResource.XHR_RESPONSE_TYPE.TEXT
@@ -283,7 +285,7 @@ export class GrpLayer extends Layer {
 		if (! meta?.image) {next(); return;}
 
 		const fn = getFn(meta.image);
-		const url = GrpLayer.#cfg.searchPath(fn, SEARCH_PATH_ARG_EXT.SPRITE);
+		const url = GrpLayer.#cfg.searchPath(fn, SEARCH_PATH_ARG_EXT.SP_GSM);
 		(new Loader)
 		.use((res2, next2)=> {
 			this.#sys.dec(res2.extension, res2.data)
@@ -413,7 +415,7 @@ export class GrpLayer extends Layer {
 		if (aImg) {aImg.push(img); return}
 		this.#hAEncImg[src] = [img];
 
-		const url2 = GrpLayer.#cfg.searchPath(src, SEARCH_PATH_ARG_EXT.SPRITE);
+		const url2 = GrpLayer.#cfg.searchPath(src, SEARCH_PATH_ARG_EXT.SP_GSM);
 		const ld2 = (new Loader)
 		.add({name: src, url: url2, xhrType: LoaderResource.XHR_RESPONSE_TYPE.BUFFER,});
 		if (GrpLayer.#sys.crypto) ld2.use((res, next)=> {