+++ test/Grammar.test.ts
@@ -234,27 +234,27 @@ it('test_let_expansion_0_cr', ()=> {
 
 	expect(len).toBe(6);
 	expect(aToken[0]).toBe('&text=1+2=int');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\n");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[3]).toBe(" あ。/");
-	ch = aToken[3].charAt(0);
+	ch = aToken[3].at(0);
 	expect(grm.testNoTxt(ch)).toBe(false);
 
 	expect(aToken[4]).toBe("\n");
-	ch = aToken[4].charAt(0);
+	ch = aToken[4].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[5]).toBe("あ。/");
-	ch = aToken[5].charAt(0);
+	ch = aToken[5].at(0);
 	expect(grm.testNoTxt(ch)).toBe(false);
 });
 it('test_let_expansion_0_cr_spc', ()=> {
@@ -264,15 +264,15 @@ it('test_let_expansion_0_cr_spc', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&text = 1 + 2 = int');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_let_expansion_0_cr_spc_var', ()=> {
@@ -282,15 +282,15 @@ it('test_let_expansion_0_cr_spc_var', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&text = hA[#ar$BB#]');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_let_expansion_1_tab', ()=> {
@@ -300,15 +300,15 @@ it('test_let_expansion_1_tab', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&text=1+2=int');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_let_expansion_1_tab_spc', ()=> {
@@ -318,15 +318,15 @@ it('test_let_expansion_1_tab_spc', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&text = 1 + 2 = int');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_let_expansion_1_tab_spc_var', ()=> {
@@ -336,15 +336,15 @@ it('test_let_expansion_1_tab_spc_var', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&text = hA[#ar$BB#]');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_let_expansion_2_spc', ()=> {
@@ -354,7 +354,7 @@ it('test_let_expansion_2_spc', ()=> {
 
 	expect(len).toBe(1);
 	expect(aToken[0]).toBe('&text=1+2=int [tcy t="!!"]');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_let_expansion_2_spc2', ()=> {
@@ -364,7 +364,7 @@ it('test_let_expansion_2_spc2', ()=> {
 
 	expect(len).toBe(1);
 	expect(aToken[0]).toBe('&text=1+2=int  [tcy t="!!"]');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_let_expansion_3_comment', ()=> {
@@ -374,15 +374,15 @@ it('test_let_expansion_3_comment', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&text=1+2=int');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('; [tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_let_expansion_3_comment2', ()=> {
@@ -392,15 +392,15 @@ it('test_let_expansion_3_comment2', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&text=1+2=int');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('; [tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_let_expansion_4_OperatorEq', ()=> {
@@ -410,15 +410,15 @@ it('test_let_expansion_4_OperatorEq', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&text=1+2 == null');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('; [tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_let_expansion_4_OperatorNotEq', ()=> {
@@ -428,15 +428,15 @@ it('test_let_expansion_4_OperatorNotEq', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&text=1+2 != null');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('; [tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_let_bug_20120321s_aira', ()=> {
@@ -446,15 +446,15 @@ it('test_let_bug_20120321s_aira', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe("&&'tex'+0 = 1+2");
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe("[tcy t='!!']");
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 
@@ -465,15 +465,15 @@ it('test_ch_expansion_0_cr', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&test&');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\n");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_ch_expansion_0_cr_spc', ()=> {
@@ -483,15 +483,15 @@ it('test_ch_expansion_0_cr_spc', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&test cr spc&');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\n");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_ch_expansion_0_cr_spc_var', ()=> {
@@ -501,15 +501,15 @@ it('test_ch_expansion_0_cr_spc_var', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&test hA[#ar$BB#]&');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\n");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_ch_expansion_1_tab', ()=> {
@@ -519,15 +519,15 @@ it('test_ch_expansion_1_tab', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&test&');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_ch_expansion_1_tab_spc', ()=> {
@@ -537,15 +537,15 @@ it('test_ch_expansion_1_tab_spc', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&test tab spc&');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_ch_expansion_1_tab_spc_var', ()=> {
@@ -555,15 +555,15 @@ it('test_ch_expansion_1_tab_spc_var', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&test hA[#ar$BB#]&');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_ch_expansion_2_spc', ()=> {
@@ -573,15 +573,15 @@ it('test_ch_expansion_2_spc', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&test&');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe(" ");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(false);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_ch_expansion_2_spc2', ()=> {
@@ -591,15 +591,15 @@ it('test_ch_expansion_2_spc2', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&test&');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("  ");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(false);
 
 	expect(aToken[2]).toBe('[tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_ch_expansion_2_comment', ()=> {
@@ -609,15 +609,15 @@ it('test_ch_expansion_2_comment', ()=> {
 
 	expect(len).toBe(3);
 	expect(aToken[0]).toBe('&test&');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\t");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('; [tcy t="!!"]');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_ch_expansion_miss_0', ()=> {
@@ -627,23 +627,23 @@ it('test_ch_expansion_miss_0', ()=> {
 
 	expect(len).toBe(5);
 	expect(aToken[0]).toBe('&te');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\n");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe("st&");
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(false);
 
 	expect(aToken[3]).toBe("\n");
-	ch = aToken[3].charAt(0);
+	ch = aToken[3].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[4]).toBe('[tcy t="!!"]');
-	ch = aToken[4].charAt(0);
+	ch = aToken[4].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_ch_expansion_miss_1', ()=> {
@@ -653,27 +653,27 @@ it('test_ch_expansion_miss_1', ()=> {
 
 	expect(len).toBe(6);
 	expect(aToken[0]).toBe('&te');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe("\n");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe('\t');
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[3]).toBe('st&');
-	ch = aToken[3].charAt(0);
+	ch = aToken[3].at(0);
 	expect(grm.testNoTxt(ch)).toBe(false);
 
 	expect(aToken[4]).toBe("\n");
-	ch = aToken[4].charAt(0);
+	ch = aToken[4].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[5]).toBe('[tcy t="!!"]');
-	ch = aToken[5].charAt(0);
+	ch = aToken[5].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 
@@ -686,19 +686,19 @@ it('test_char2macro_0', ()=> {
 
 	expect(len).toBe(4);
 	expect(aToken[0]).toBe('[tcy t="!!"]');
-	ch = aToken[0].charAt(0);
+	ch = aToken[0].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[1]).toBe(" あ。/");
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(false);
 
 	expect(aToken[2]).toBe("\n");
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[3]).toBe("あ。/");
-	ch = aToken[3].charAt(0);
+	ch = aToken[3].at(0);
 	expect(grm.testNoTxt(ch)).toBe(false);
 });
 
@@ -1061,11 +1061,11 @@ it('test_let_ml_1', ()=> {
 	{"fn": "../bgimage/yun_1317.jpg").toBe("dt": "2018/12/18 11:44").toBe("txt": "　ああ、桜の樹の下には屍体が埋まつてゐる！"},
 ]
 `);
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe(`[endlet_ml]`);
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_let_ml_2', ()=> {
@@ -1099,11 +1099,11 @@ it('test_let_ml_2', ()=> {
 	}
 }
 `);
-	ch = aToken[1].charAt(0);
+	ch = aToken[1].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 
 	expect(aToken[2]).toBe(`[endlet_ml]`);
-	ch = aToken[2].charAt(0);
+	ch = aToken[2].at(0);
 	expect(grm.testNoTxt(ch)).toBe(true);
 });
 it('test_let_ml_2022/10/15_0', ()=> {